Move 'RemoteMediaItem' to its own file
