+++ MovieListFramework/Media API/MediaItemsMapper.swift
@@ -7,24 +7,6 @@
 
 import Foundation
 
-internal struct RemoteMediaItem: Decodable {
-    internal let adult: Bool
-    internal let backdrop_path: String?
-    internal let genre_ids: [Int]
-    internal let id: Int
-    internal let media_type: String?
-    internal let original_language: String
-    internal let original_title: String
-    internal let overview: String
-    internal let popularity: Double
-    internal let poster_path: String?
-    internal let release_date: String
-    internal let title: String
-    internal let video: Bool
-    internal let vote_average: Double
-    internal let vote_count: Int
-}
-
 internal class MediaItemsMapper {
     private struct Root: Decodable {
         let results: [RemoteMediaItem]
